Model is now a Base Class
